requirements.txt
streamlit
azure-ai-projects
azure-identity
streamlit-option-menu
